Multi-file schema, allow multiple diagrams, addons
Major revision
Fixes #1
Fixes #5

Starts #6 (though I may consider this one ongoing)

NOTE: Changes are mirrored between `docs/` and `versioned_docs/version-Live` Contents of these directories should be identical. Only the "Next" version in `docs/` is previewable. Other changes will apply to the Live version once these changes are merged to `main`.

-- Split schema into multiple files --

Defined sub-schemas that are composed with JSON Schema references.
https://json-schema.org/understanding-json-schema/structuring#dollarref

- Causal-Decision-Model.json
	- Base schema for the overall model
- DI-Diagram.json
	- Defines the diagram objects that can be associated with a model. These will be your CDDs, made up of smaller graphical elements (boxes, lines, text, etc).
- DI-Evaluatable.json
	- Defines the evaluatable models or model fragments that provide functionality for evaluating or simulating the CDM. Similar to diagrams, these are made up of smaller evaluatable elements (scripts, API Calls, etc).
- Causal-Type.json
	- Defines the basic types commonly used in CDDs, while providing a mechanism for custom types (named "CUSTOM_[something]) if desired. These are used in diagram and evaluatable elements.
- DI-Addons.json
	- Defines the basic structure for all addon objects. These need UUIDs for owner and addon, and otherwise provide a basic bucket for storing any other JSON data, with a little prompting for having addon creators define sub-schemas for their addons.
- DI-Asset.json
	- Defines the basic metadata meant to be associated with any object that might be retrieved from this CDM. That includes the CDM itself, any diagrams or evaluatables, addons, and individual evaluatable or diagram elements.
- UUID.json
	- Defines a basic regex pattern for UUIDs adherent to UUID v4

-- Allow multiple diagrams and evaluatable models --

This solves the issue where often a given CDM will be composed of multiple model fragments, and may have multiple diagrams associated with it. CDMs are now cleanly split into diagrams and evaluatable models, allowing for multiple of each, with explicit relationships between diagram and evaluatable elements allowed but not required.

-- Begin defining graphical vocabulary for CDDs --

These are defined in DI-Diagram.json, and include the following for now:
- Point
- Line
- Box
- Ellipse
- Text
- (Custom)

I suspect these will still need tweaking, and this list may still see some additions, but I'm leaving it here for now. We'll likely work out what I've missed when we try to use this schema to construct CDDs.

Related to this, there are still the following _evaluatable_ elements defined in DI-Evaluatable.json:
- Script
- API Call
- Binary Base64
- (Custom)

I'd like these two lists to be an ongoing area of discussion and refinement.

-- Changes to docs --

Splitting the schema into separate files has presented an opportunity to document these schema fragments separately, but this is still on the TODO.

I've updated index.md with code to help resolve references to local files, per this:
https://jy95.github.io/docusaurus-json-schema-plugin/docs/demo-viewer/refs/localFile

In the future, I'd like to break the docs into multiple pages focused on viewing each sub-schema on its own, so this reference resolver code will need to be made more reuseable.