API Docs: Justified endpoints with links to Roles
Added some more motivating documentation to contextualize endpoints within the existing OpenDI Roles and User Stories.
Added a "Roles" section to the doc page associated with each endpoint category, with a list of roles that are relevant for endpoints in that category.

Intent is to facilitate/contextualize future discussion and development of these endpoints.