Add PUT + POST operations for /v0/models endpoint
Updated api source JSON to add descriptions for PUT (update existing model) and POST (create new model) operations on the `/v0/models` endpoint.

Also, brought up-to-date OpenAPI version of the docs over to the Live version. Intent is to finally merge this to main, so that `opendi.org/api-specification` shows the latest API spec.